Fixes config vars reading.
To read config vars from build packs, it is necessary to query them off
the env_dir parameter passed to compiler. Each environment variable is a
file, where the file name is the env key and its contents, the value.

Removing a file from the directory would make it so the environment
variable will not be available at run scope. So, the path to retype in
the repository and also the license key variables are consumed by the
compiler script and removed so that they aren't forwarded.