Builds lighttpd mime types off mime-db database.
Introduced the contrib/fetch-lighttpd-mimetypes.js script to handle
conversion of the mime-db database:
- Fetches fresh version of the json database off the repository's
  'master' branch.
- Resolves some extension conflicts using the (hopefully) best choice.
- Sorts the mime types by the extension, alphabetically.
- Outputs in lighttpd format, wrapping with the actual
  'mimetypes.assign' directive.

The output of the script can then be redirected into the
include/lightppd-mime-types.conf file to update it.

Then, changed the compile script to copy over and include the mime types
config file into the default config file.

Running the fetch script every at build time would ensure the list is
always fresh but would require full NodeJS implementation, or building a
C# console app to do the conversion. Given the mime types list is not
so frequently updated, decided to keep it simple and manually update the
file once in a while.